Prepare for elliptic
