+++ lib/js/ecpoint.js
