+++ lib/js/ecpointg.js
