+++ lib/js/rfc6979/index.js
